project-root/services/sales_service/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
jinja2
weasyprint
PyJWT
kafka-python==2.0.2

